Move MallocMCBuffer to PMacc
