+++ src/libPMacc/include/particles/memory/buffers/MallocMCBuffer.hpp
@@ -1,36 +1,35 @@
 /**
  * Copyright 2015 Rene Widera
  *
- * This file is part of PIConGPU.
+ * This file is part of libPMacc.
  *
- * PIConGPU is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
+ * libPMacc is free software: you can redistribute it and/or modify
+ * it under the terms of either the GNU General Public License or
+ * the GNU Lesser General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
  *
- * PIConGPU is distributed in the hope that it will be useful,
+ * libPMacc is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * GNU General Public License and the GNU Lesser General Public License
+ * for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with PIConGPU.
+ * and the GNU Lesser General Public License along with libPMacc.
  * If not, see <http://www.gnu.org/licenses/>.
  */
-
 
 #pragma once
 
 
 #include "dataManagement/ISimulationData.hpp"
-#include "simulation_defines.hpp"
 
 #include "mallocMC/mallocMC.hpp"
 #include <string>
 
-namespace picongpu
+namespace PMacc
 {
-    using namespace PMacc;
 
     class MallocMCBuffer : public ISimulationData
     {
@@ -66,6 +65,6 @@ namespace picongpu
     };
 
 
-} // namespace picongpu
+} // namespace PMacc
 
-#include "particles/MallocMCBuffer.tpp"
+#include "particles/memory/buffers/MallocMCBuffer.tpp"
